requirements.txt
aiogram==2.21
psycopg2~=2.9.10
apscheduler==3.8.1
qrcode==7.3.1
flask==2.0.2
google-api-python-client==2.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6

asyncio~=3.4.3